requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
lightgbm>=4.0
xgboost>=2.0
catboost>=1.2
geopandas>=0.14
rasterio>=1.3
pyyaml>=6.0
matplotlib>=3.7
shap>=0.44
jupyter>=1.0
ipykernel>=6.0
